Reduced functionality to 32bit.

+++ include/generic-reveng/analysis/data_section.hpp
@@ -6,7 +6,7 @@ namespace grev
 {
     struct data_section
     {
-        std::uint64_t address;
+        std::uint32_t address;
         std::u8string_view data;
     };
 }
