+++ source/loading/pe/pe_program.hpp
@@ -7,7 +7,7 @@ namespace grev
     class pe_program : public program
     {
         machine_architecture architecture_;
-        std::uint64_t start_address_;
+        std::uint32_t start_address_;
 
         std::set<address_space_segment, address_space_segment::exclusive_address_order> segments_;
 
@@ -16,7 +16,7 @@ namespace grev
         explicit pe_program(std::u8string data);
 
         machine_architecture architecture() const override;
-        std::uint64_t start_address() const override;
+        std::uint32_t start_address() const override;
 
     protected:
